Made changes provided by self-review exercise
